[fix] apply_softcap

+++ csrc/flash_attn/src/flash_fwd_kernel.h
@@ -25,7 +25,7 @@ using namespace cute;
 template <typename Engine, typename Layout>
 __forceinline__ __device__ void apply_softcap(Tensor<Engine, Layout> &tensor, const float softcap){
     #pragma unroll
-    Tensor softcap_att = make_tensor(softcap, Float<1>{});
+    Tensor softcap_att = make_tensor(softcap, Int<1>{});
     for (int i = 0; i < size(tensor); ++i) {
         tensor(i) = cutlass::fast_tanh(tensor(i) * softcap);
     }
